portalnetwork: rename receiptManager
