+++ packages/portalnetwork/src/subprotocols/index.ts
@@ -1,5 +1,5 @@
 export * from './history/index.js'
 export * from './types.js'
 export * from './protocol.js'
-export * from './receipt.js'
+export * from './receiptManager.js'
 export * from './nodeLookup.js'
